Click on 'delete' button will 'unregister' for browsable Data Providers.
git-svn-id: svn+ssh://huia.bic.mni.mcgill.ca/home/cbrain/cbrain_repo/trunk@515 bcc1f7e1-216c-4ce3-888e-c5a8bb225047

+++ BrainPortal/app/controllers/userfiles_controller.rb
@@ -410,8 +410,14 @@ class UserfilesController < ApplicationController
       when "delete"
         Userfile.find_accessible_by_user(filelist, current_user, :access_requested => :write).each do |userfile|
           basename = userfile.name
-          userfile.destroy
-          flash[:notice] += "File #{basename} deleted.\n"
+          if userfile.data_provider.is_browsable?
+            userfile.destroy_log rescue true
+            userfile.delete
+            flash[:notice] += "File '#{basename}' unregistered.\n"
+          else
+            userfile.destroy
+            flash[:notice] += "File '#{basename}' deleted.\n"
+          end
         end
 
       when "download"
